requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
dotenv==0.9.9
fastapi==0.119.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
mysql-connector==2.2.9
mysql-connector-python==9.5.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
